// src/app/model/Contado.ts
export class Contado {
    id: number;
    numero_contrato: number;
    objeto_contrato: string;
    nombreContratista: string;
    ruc_contratista: number;
    numOrden: number;
    cantidad: number;
    precioUnitario: number;
    subtotal: number;
    iva: number;
    otrosImpuestos: number;
    total: number;
    cod_presupuestario: number;
    descripcionpresupuestaria: string;
    valorCert: number;
    afectuar: number;
    fechaHoraIni: string;
    email: string;
    descripcion: string;
    obtencion_iva: number;
    obtencion_otro_impuesto: number;
    mes: string;

}